Buttons rework
Moved non-KC button processing to `buttons.c``; events are now stored
internally in a `k_pipe` object.
Other parts of the code can retrieve button press/release events via
buttons_get(), with or without a timeout, with or without a filter.

Adapted snake to use the new system; now it is possible to plan a snake
path in advance.